+++ game-aska/CHANGELOG.md
@@ -4,8 +4,14 @@
 
 - None
 
+## [0.1.2] - 2026-03-16
+
+- Added: Notification and button to download latest BepInExConfigurationManager
+
 ## [0.1.1] - 2026-03-15
 
+- Changed: BepInEx is now the downloaded without a popup choice, since it has become the standard for modding. MelonLoader can still be installed manually and is supported
+- Added: Button to browse to download the latest BepInEx BE version
 - Fixed: Bump BepInEx BE version to 755 and fix URL
 - Added: .NET 6 installation check and notification for MelonLoader
 - Added: Automatic detection of BepInEx patchers